pyproject.toml
[project]
authors = [{name = "Nhat-Minh Nguyen"}]
description = "Spotify Wrapped, but more sus."
readme = "README.md"
license = {file = "LICENCE"}
requires-python = ">= 3.8"
dependencies = [
    "Flask-Session==0.5.0",    # fix trouble migrating Airflow db at local
    "apache-airflow==2.8.1",
    "apache-airflow-providers-apache-spark==4.7.1",
    "Faker==23.2.1"
]

[project.optional-dependencies]
dev = ["pytest==8.0.1", "mypy==1.8.0", "mypy-extensions==1.0.0", "black"]

[tool.pytest.ini_options]
pythonpath = "src/dags src"
testpaths = "tests"
addopts = "-ra -q -p no:cacheprovider"

[tool.pytest_env]
PYARROW_IGNORE_TIMEZONE = 1     # to suppress warning when testing PySpark jobs